now the program can work with rational numbers

+++ Assignment2/b/polish_math.cpp
@@ -9,10 +9,10 @@
 #include <sstream>
 
 void PolishMathing(std::string expression){
-    std::stack<int> num_stack; // create stack, where we will store the last results and numbers
+    std::stack<double> num_stack; // create stack, where we will store the last results and numbers
     std::stringstream stream(expression); // create a pointer that will traverse the entered string(expression)
     std::string value;
-    int el1, el2; // create two variables that will be used for calculations
+    double el1, el2; // create two variables that will be used for calculations
     while(stream>>value){
         if (value=="+"){
             el1 = num_stack.top();
@@ -43,7 +43,7 @@ void PolishMathing(std::string expression){
             num_stack.push(el2);
         }
         else {
-            num_stack.push(stoi(value));
+            num_stack.push(stod(value));
         }
     }
     std::cout << "result: " << num_stack.top() << std::endl;
